linear approach

+++ libccar.h
@@ -366,7 +366,7 @@ static float lcc_clamp(float v, float lo, float hi) {
 }
 
 /* clamp v between [lo,hi] */
-static float lcc_iclamp(int v, int lo, int hi) {
+static int lcc_iclamp(int v, int lo, int hi) {
   return v < lo ? lo : (v > hi ? hi : v);
 }
